requirements.txt
absl-py==2.1.0
asttokens==3.0.0
backcall==0.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.0
executing==2.2.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
labelImg==1.8.6
lxml==5.3.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==1.4.4
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.9.3
seaborn==0.12.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
